+++ src/MasterTool/Features/GodMode/DamagePatches.cs
@@ -4,7 +4,6 @@ using EFT;
 using EFT.HealthSystem;
 using HarmonyLib;
 using MasterTool.Config;
-using MasterTool.Features.CodMode;
 using MasterTool.Plugin;
 
 namespace MasterTool.Features.GodMode
@@ -184,8 +183,6 @@ namespace MasterTool.Features.GodMode
                         damage = Math.Max(0f, currentHealth.Current - 3f);
                     }
                 }
-
-                CodModeFeature.NotifyDamage();
 
                 return true;
             }